change state for each reservation
