+++ front_end/src/components/history/history.css
@@ -218,7 +218,7 @@
   text-align: center;
 }
 
-.center {
+.centerChangeReservation {
   text-align: center;
   margin: auto;
   width: 60%;
